arm64: configs: Enable inline storage encryption
Most Qualcomm platforms released in the past decade feature an
Inline Crypto Engine, used for accelerating crypto operations on
eMMC and UFS media.

Inline encryption is widely used on Android and can be utilized
with any other userland. Enable the required configs to support
both the Linux infrastructure for it and the Qualcomm IP block.

Signed-off-by: Gaurav Kashyap <gaurav.kashyap@oss.qualcomm.com>
Signed-off-by: Jingyi Wang <jingyi.wang@oss.qualcomm.com>