media: qcom: camss: csiphy: Add support for v2.4.0 two-phase CSIPHY
Add more detailed resource information for CSIPHY devices in the camss
driver along with the support for v2.4.0 in the 2 phase CSIPHY driver
that is responsible for the PHY lane register configuration, module
reset and interrupt handling.

This change adds 'cmn_status_offset' variable in 'csidphy_device_regs'
structure. It helps adapt the offset to the common status registers that
is different in v2.4.0 from others.

Signed-off-by: Hangxiang Ma <hangxiang.ma@oss.qualcomm.com>
Signed-off-by: Jingyi Wang <jingyi.wang@oss.qualcomm.com>